Chore: Move core tests into directory
