+++ Test/CMakeLists.txt
@@ -3,7 +3,7 @@ CMAKE_MINIMUM_REQUIRED(VERSION 3.26)
 PROJECT(${TEST_PROJECT_NAME} VERSION 1.0.0 LANGUAGES CXX)
 
 SET(ENTRY_POINT
-		Source/Main.cpp)
+		Source/Core/Main.cpp)
 
 ADD_EXECUTABLE(${TEST_PROJECT_NAME} ${ENTRY_POINT})
 TARGET_INCLUDE_DIRECTORIES(${PROJECT_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/Include/)
