+++ Test/Include/CMakeLists.txt
@@ -1,6 +1,7 @@
 CMAKE_MINIMUM_REQUIRED(VERSION 3.26)
 
-SET(HEADER_FILES
-		FakeMazeGenerator.hpp)
+SET(HEADER_FILES)
 
 TARGET_SOURCES(${PROJECT_NAME} PRIVATE ${HEADER_FILES})
+
+ADD_SUBDIRECTORY(Core)
