+++ Test/Include/Core/CMakeLists.txt
@@ -1,0 +1,6 @@
+CMAKE_MINIMUM_REQUIRED(VERSION 3.26)
+
+SET(HEADER_FILES
+		FakeMazeGenerator.hpp)
+
+TARGET_SOURCES(${PROJECT_NAME} PRIVATE ${HEADER_FILES})
